Update and rename 01-starting-with-data.Rmd to 02-starting-with-data.Rmd
removed items from statistics onward. this allows it to segue into working with subsetting data, which I think we should cover next as a deep dive into working with data frames, vectors, matrices, etc. 
I also renamed it 02... so it can match properly.  we'll need to delete the extra 01 in the `_site` and `episodes` folders (folder in first, file in second). that was the only I could rename everything and get them to render and advance properly in the lesson locally.