Update and rename 05-data-subsetting.Rmd to 03-data-subsetting.Rmd
I changed a lot of the examples to give a little more meaningful variable names, even if it is generic `sample`. 

Also, I think this episode should go right after 02, since it covers data frames and subsetting data more in depth.  I think we talked about it.

I added the `ls` part here.  I think there's enough data in the environment to make that worthwhile.  I stopped right before `challenge 3`, which I think doesn't make sense in terms of the way we are working the lesson.  

I also fixed the line lengths so they don't go over 100(ish) characters.  it's current practice when dealing with git, so it's easier to see where the changes are, especially if there's only a small change.